Updated enable SSL instructions in docs
Instead of using browser (when running Grafana) to connect to InfluxDB
(that requires connection to be on https too due to security
requirements, and to do so use Apache proxy from 10443 -> 8086), we are
now using grafana-server proxy (so it can just use plain http to
localhost:87086).
That saves DNS lookup, HTTPS transport layer, Apache proxy to localhost
etc.
This speeds up InfluxDB data source and also allows to work when high
ports are blocke (like on some WiFi networks).

Signed-off-by: Lukasz Gryglicki <lukaszgryglicki@o2.pl>